Update README.md
This was another special “one-off” report that one member of the Executive Committee wanted answered.  The question was “How does our Accounting Coding costs compare to actual EDL delays during production?”    
This was answered by first contacting and understanding how they code accounting costs for each vessel by what Category, then our maintenance databases, and then the EDL database.  This is about as linking the relationships between that I had at the time, so it seems very elementary but that is what I had to work with.  After that, it was a matter of data-mining and graphing my analysis into this report.  I do not have the original file for this to show my work on this specific report, only my final PowerPoint and Adobe PDF.